Updates target framework and dependencies
Updates the target framework for all projects to net10.0.

Also updates the versions of nuget packages used in the projects.
Removes the TeachTrack.Model project and moves its content to TeachTrack.Core

These updates ensure compatibility with the latest .NET runtime and libraries and improve the overall stability of the application.