Implements Academic Standing entity.
Creates the `AcademicStanding` entity with name and description properties, including validation for name.

Removes the previous implementation of Academic Standing.

Adds EF Core configuration for the new AcademicStanding entity, including table and column mappings, primary key, unique index on the name, and default ID generation.

Adds a reference to the Core project in the Model project.